relocate ParticipantDataProxy
