+++ apps/pipelines/tests/test_runnable_builder.py
@@ -12,7 +12,6 @@ from apps.experiments.models import ParticipantData
 from apps.pipelines.exceptions import PipelineBuildError, PipelineNodeBuildError
 from apps.pipelines.logging import LoggingCallbackHandler
 from apps.pipelines.nodes.base import PipelineState
-from apps.pipelines.nodes.helpers import ParticipantDataProxy
 from apps.pipelines.nodes.nodes import EndNode, StartNode, StaticRouterNode
 from apps.pipelines.tests.utils import (
     assistant_node,
@@ -32,6 +31,7 @@ from apps.pipelines.tests.utils import (
     state_key_router_node,
 )
 from apps.service_providers.llm_service.history_managers import PipelineHistoryManager
+from apps.service_providers.llm_service.prompt_context import ParticipantDataProxy
 from apps.service_providers.llm_service.runnables import ChainOutput
 from apps.utils.factories.assistants import OpenAiAssistantFactory
 from apps.utils.factories.experiment import (
