+++ apps/service_providers/llm_service/prompt_context.py
@@ -1,8 +1,9 @@
-from typing import Any
+from typing import Any, Self
 
 from django.utils import timezone
 
 from apps.channels.models import ChannelPlatform
+from apps.experiments.models import ParticipantData
 from apps.utils.time import pretty_date
 
 
@@ -134,3 +135,61 @@ class SafeAccessWrapper(dict):
 
 
 EMPTY = SafeAccessWrapper(None)
+
+
+class ParticipantDataProxy:
+    """Allows multiple access without needing to re-fetch from the DB"""
+
+    @classmethod
+    def from_state(cls, pipeline_state) -> Self:
+        # using `.get` here for the sake of tests. In practice the session should always be present
+        return cls(pipeline_state.get("experiment_session"))
+
+    def __init__(self, experiment_session):
+        self.session = experiment_session
+        self._participant_data = None
+
+    def _get_db_object(self):
+        if not self._participant_data:
+            self._participant_data, _ = ParticipantData.objects.get_or_create(
+                participant_id=self.session.participant_id,
+                experiment_id=self.session.experiment_id,
+                team_id=self.session.experiment.team_id,
+            )
+        return self._participant_data
+
+    def get(self):
+        data = self._get_db_object().data
+        return self.session.participant.global_data | data
+
+    def set(self, data):
+        participant_data = self._get_db_object()
+        participant_data.data = data
+        participant_data.save(update_fields=["data"])
+
+        self.session.participant.update_name_from_data(data)
+
+    def get_schedules(self):
+        """
+        Returns all active scheduled messages for the participant in the current experiment session.
+        """
+        from apps.events.models import ScheduledMessage
+
+        experiment = self.session.experiment_id
+        participant = self.session.participant_id
+        team = self.session.experiment.team
+        messages = (
+            ScheduledMessage.objects.filter(
+                experiment_id=experiment,
+                participant_id=participant,
+                team=team,
+                is_complete=False,
+                cancelled_at=None,
+            )
+            .select_related("action")
+            .order_by("created_at")
+        )
+        scheduled_messages = []
+        for message in messages:
+            scheduled_messages.append(message.as_dict())
+        return scheduled_messages
